Final database created

+++ HrApp.Domain/Entities/Offer.cs
@@ -4,7 +4,7 @@ public class Offer
 {
     public Guid Id { get; set; }
     public string PositionName { get; set; } = default!;
-    public decimal Salary { get; set; }
+    public float Salary { get; set; }
     public string Descritpion { get; set; } = default!;
     public DateOnly AddDate { get; set; }
     public List<Application> Applications { get; set; } = default!;
